+++ packages/apollo/src/hooks/useLazyQuery.ts
@@ -1,0 +1,3 @@
+import { useLazyQuery } from '@apollo/react-hooks';
+
+export default useLazyQuery;