+++ packages/apollo/src/hooks/useMutation.ts
@@ -1,0 +1,3 @@
+import { useMutation } from '@apollo/react-hooks';
+
+export default useMutation;