+++ packages/core/src/components/Divider.story.tsx
@@ -6,4 +6,5 @@ storiesOf('Core/Divider', module)
   .addParameters({
     inspectComponents: [Divider],
   })
-  .add('Standard divider.', () => <Divider />);
+  .add('Standard divider.', () => <Divider />)
+  .add('Short divider.', () => <Divider short />);