+++ packages/core/src/index.ts
@@ -16,6 +16,7 @@ import {
   TranslateParams,
   TranslateProps,
   Theme,
+  TranslateOptions,
 } from './types';
 
 export type Settings = {
@@ -133,13 +134,18 @@ class Core {
     return this.settings.defaultTimezone || DEFAULT_TIMEZONE;
   }
 
-  translate = (phrase: string, params: TranslateParams, context: string) => {
+  translate = (
+    phrase: string,
+    baseParams?: TranslateParams | null,
+    options?: string | TranslateOptions,
+  ) => {
     const { translator } = this.settings;
 
     if (translator) {
-      return translator(phrase, params, context);
+      return translator(phrase, baseParams, options);
     }
 
+    const params = baseParams || {};
     let message = phrase;
 
     if (phrase.includes('||||')) {