+++ packages/forms/test/components/Form/Input.test.tsx
@@ -1,29 +1,32 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { mount } from 'enzyme';
 import BaseInput from '@airbnb/lunar/lib/components/Input';
-import { unwrapHOCs } from '@airbnb/lunar-test-utils';
 import Input from '../../../src/components/Form/Input';
 import { toString } from '../../../src/helpers';
+import { Context } from '../../../src/types';
+import { WrappingFormComponent, createFormContext } from '../../utils';
 
 describe('<Input />', () => {
-  const form = {
-    change() {},
-    getState: () => ({} as any),
-    register: jest.fn(),
-  };
+  let context: Context;
+
+  beforeEach(() => {
+    context = createFormContext();
+  });
 
   it('connects to the form', () => {
-    const wrapper = unwrapHOCs(
-      shallow(<Input label="Label" name="foo" defaultValue="bar" validator={() => {}} />),
-      'FormInput',
-      form,
+    const wrapper = mount(
+      <Input label="Label" name="foo" defaultValue="bar" validator={() => {}} />,
+      {
+        wrappingComponent: WrappingFormComponent,
+        wrappingComponentProps: { context },
+      },
     );
 
-    expect(form.register).toHaveBeenCalledWith(
+    expect(context.register).toHaveBeenCalledWith(
       expect.objectContaining({ name: 'foo', defaultValue: 'bar', parse: toString }),
       expect.anything(),
     );
 
-    expect(wrapper.shallow().find(BaseInput)).toHaveLength(1);
+    expect(wrapper.find(BaseInput)).toHaveLength(1);
   });
 });