+++ packages/forms/test/components/Form/Select.test.tsx
@@ -1,35 +1,36 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { mount } from 'enzyme';
 import BaseSelect from '@airbnb/lunar/lib/components/Select';
-import { unwrapHOCs } from '@airbnb/lunar-test-utils';
 import Select from '../../../src/components/Form/Select';
 import { toString } from '../../../src/helpers';
+import { Context } from '../../../src/types';
+import { WrappingFormComponent, createFormContext } from '../../utils';
 
 describe('<Select />', () => {
-  const form = {
-    change() {},
-    getState: () => ({} as any),
-    register: jest.fn(),
-  };
+  let context: Context;
+
+  beforeEach(() => {
+    context = createFormContext();
+  });
 
   it('connects to the form', () => {
-    const wrapper = unwrapHOCs(
-      shallow(
-        <Select label="Label" name="foo" defaultValue="bar" validator={() => {}}>
-          <option value="foo">Foo</option>
-          <option value="bar">Bar</option>
-          <option value="baz">Baz</option>
-        </Select>,
-      ),
-      'FormSelect',
-      form,
+    const wrapper = mount(
+      <Select label="Label" name="foo" defaultValue="bar" validator={() => {}}>
+        <option value="foo">Foo</option>
+        <option value="bar">Bar</option>
+        <option value="baz">Baz</option>
+      </Select>,
+      {
+        wrappingComponent: WrappingFormComponent,
+        wrappingComponentProps: { context },
+      },
     );
 
-    expect(form.register).toHaveBeenCalledWith(
+    expect(context.register).toHaveBeenCalledWith(
       expect.objectContaining({ name: 'foo', defaultValue: 'bar', parse: toString }),
       expect.anything(),
     );
 
-    expect(wrapper.shallow().find(BaseSelect)).toHaveLength(1);
+    expect(wrapper.find(BaseSelect)).toHaveLength(1);
   });
 });